Teste na classe UserController

+++ src/main/java/com/curso/pdv/controller/UserController.java
@@ -3,6 +3,7 @@ package com.curso.pdv.controller;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -14,7 +15,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.curso.pdv.dto.ResponseDTO;
 import com.curso.pdv.entity.User;
+import com.curso.pdv.exceptions.NoItemException;
 import com.curso.pdv.repository.UserRepositry;
 
 @Controller
@@ -25,38 +28,49 @@ public class UserController {
     private UserRepositry userRepositry;
 
     @GetMapping
-    public ResponseEntity getAll(){
+    public ResponseEntity<?> getAll(){
         return new ResponseEntity<>(userRepositry.findAll(), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity post(@RequestBody User user){
+    public ResponseEntity<?> post(@RequestBody User user){
         try{
             user.setEnable(true);
             return new ResponseEntity<>(userRepositry.save(user),HttpStatus.CREATED);
         }catch(Exception error){
-            return new ResponseEntity<>(error.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(new ResponseDTO<>(error.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PutMapping
-    public ResponseEntity put(@RequestBody User user){
+    public ResponseEntity<?> put(@RequestBody User user){
         Optional<User> userToEdit = userRepositry.findById(user.getId());
         if(userToEdit.isPresent()){
-            userRepositry.save(user);
-            return new ResponseEntity<>(user,HttpStatus.OK);
+            try{
+                userRepositry.save(user);
+                return new ResponseEntity<>(user,HttpStatus.OK);
+            }catch(NoItemException error){
+                return new ResponseEntity<>(new ResponseDTO<>(error.getMessage()),HttpStatus.BAD_REQUEST);
+            }catch(Exception error){
+                return new ResponseEntity<>(new ResponseDTO<>(error.getMessage()),HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         }
-        
-        return ResponseEntity.notFound().build();
+        return new ResponseEntity<>(new ResponseDTO<>("Usuário nao encontrado!"),HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity delete(@PathVariable long id){
-        try{
-            userRepositry.deleteById(id);
-            return new ResponseEntity<>("Usuário removido com sucesso", HttpStatus.OK);
-        }catch(Exception error){
-            return new ResponseEntity<>(error.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    public ResponseEntity<?> delete(@PathVariable long id){
+        Optional<User> userToEdit = userRepositry.findById(id);
+        if(userToEdit.isPresent()){
+            try{
+                userRepositry.deleteById(id);
+                return new ResponseEntity<>(new ResponseDTO<>("Usuário removido com sucesso"), HttpStatus.OK);
+            }catch(EmptyResultDataAccessException error){
+                return new ResponseEntity<>(new ResponseDTO<>(error.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+            }catch(Exception error){
+                return new ResponseEntity<>(new ResponseDTO<>(error.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         }
+        return new ResponseEntity<>(new ResponseDTO<>("Usuário nao encontrado!"),HttpStatus.NOT_FOUND);    
     }
 }
